% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obart.R
\name{obart}
\alias{obart}
\title{obart}
\usage{
obart(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  cont = FALSE,
  rm.const = TRUE,
  k = 2,
  power = 2,
  base = 0.95,
  rho = NULL,
  ntree = 200L,
  numcut = 100L,
  ndpost = 1000L,
  nskip = 100L,
  oldgamma,
  oldz,
  printevery = 100,
  offset
)
}
\arguments{
\item{x.train}{Explanatory variables for training data.}

\item{y.train}{Binary dependent variable for training (in sample) data.}

\item{x.test}{Explanatory variables for test data.}

\item{xinfo}{A list (matrix) where the items (rows) are the covariates and 
the contents of the items (columns) are the cutpoints.}

\item{usequants}{If usequants = FALSE, then the cutpoints in xinfo are generated uniformly; 
otherwise, if TRUE, uniform quantiles are used for the cutpoints.}

\item{cont}{Whether or not to assume all variables are continuous.}

\item{rm.const}{Whether or not to remove constant variables.}

\item{k}{The number of prior standard deviations \eqn{f(x)} is away from \eqn{+/-3}. 
The bigger k is, the more conservative the fitting will be.}

\item{power}{Power parameter for tree prior.}

\item{base}{Base parameter for tree prior.}

\item{rho}{Sparse parameter: typically \eqn{\rho=p} where \eqn{p} is the number of 
covariates under consideration.}

\item{ntree}{The number of trees in the sum.}

\item{numcut}{The number of possible values of c.}

\item{ndpost}{The number of posterior draws returned.}

\item{nskip}{Number of MCMC iterations to be treated as burn in.}

\item{oldgamma}{Initial value of cutoff values.}

\item{oldz}{Initial value of latent continuous variables.}

\item{printevery}{As the MCMC runs, a message is printed every printevery draws.}
}
\description{
obart
}
