% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.obart.R
\name{predict.obart}
\alias{predict.obart}
\title{predict.obart}
\usage{
\method{predict}{obart}(object, newdata, mc.cores = 1, openmp = (mc.cores.openmp() > 0), ...)
}
\arguments{
\item{mc.cores}{Description of the second parameter.}

\item{x.train}{Description of the first parameter.}

\item{y.train}{Description of the second parameter.}

\item{x.test}{Description of the second parameter.}

\item{xinfo}{Description of the second parameter.}

\item{usequants}{Description of the second parameter.}

\item{cont}{Description of the second parameter.}

\item{rm.const}{Description of the second parameter.}

\item{k}{Description of the second parameter.}

\item{power}{Description of the second parameter.}

\item{base}{Description of the second parameter.}

\item{rho}{Description of the second parameter.}

\item{ntree}{Description of the second parameter.}

\item{numcut}{Description of the second parameter.}

\item{ndpost}{Description of the second parameter.}

\item{nskip}{Description of the second parameter.}

\item{oldgamma}{Description of the second parameter.}

\item{oldz}{Description of the second parameter.}

\item{nice}{Description of the second parameter.}

\item{seed}{Description of the second parameter.}
}
\value{
What the function returns.
}
\description{
predict.obart
}
